take comments into account
