+++ fhirpipe/utils.py
@@ -1,5 +1,4 @@
 import re
-from collections import defaultdict
 
 
 def build_col_name(table, column, owner=None):
@@ -32,33 +31,3 @@ def get_table_name(name):
             TABLE.COLUMN -> TABLE
     """
     return re.search(column_pattern, name).group(1)
-
-
-def dict_concat(dict_1, dict_2):
-    for key, val in dict_2.items():
-        dict_1[key] += val
-
-
-def build_join_graph(joins):
-    """
-    Transform a join info into SQL fragments and parse the graph of join dependency
-    Input:
-        {("<owner>.<table>.<col>", "<owner>.<join_table>.<join_col>"), ... }
-    Return:
-        {
-            "<table>": ["<join_table 1>", "<join_table 2>", ...],
-            ...
-        }
-    """
-    graph = defaultdict(list)
-    for join in joins:
-        join_source, join_target = join
-
-        # Get table names
-        target_table = get_table_name(join_target)
-        source_table = get_table_name(join_source)
-
-        # Add the join in the join_graph
-        graph[source_table].append(target_table)
-
-    return graph
