+++ requirements.txt
@@ -1,5 +1,5 @@
-fhirstore
-cleaning-scripts
+fhirstore==0.1.0
+cleaning-scripts==0.1
 cx_Oracle==7.1.2
 psycopg2==2.8.3
 tqdm==4.36.1
